Only kill custom backends if they exist. Otherwise we kill ourself!
git-svn-id: https://svn.callweaver.org/callweaver/trunk@3892 e56c2a57-502e-0410-a6e2-86b2f1ef08c6

+++ res/res_musiconhold.c
@@ -1160,7 +1160,8 @@ static void moh_killall(void)
 	struct mohclass *class;
 
 	for (class = mohclasses; class; class = class->next)
-		kill(class->pid, SIGKILL);
+		if (class->pid)
+			kill(class->pid, SIGKILL);
 }
 
 static struct opbx_atexit moh_atexit = {
